Extend Workflow 115 for remaining secondary errors
