+++ scripts/115_apply_a52xq_secondary_compat.py
@@ -1,0 +1,58 @@
+#!/usr/bin/env python3
+from __future__ import annotations
+import base64
+import zlib
+
+_PAYLOAD = (
+    'c-qw)Yj5L5?sNPrrdVJlrG6y4?H%AW0n**H4VrBB@@P>c=|GW2j@EkZjFiU(x!-'
+    '=|JnWGy+uAH}?JYL8G$J`9hvbkW`qek1O7PK5oQ+s^M~cVtHqYMnBA%y&(x@sc&L|~uTI9SWewO8>U&eVRdOb77Cxy=iGxwiFo|$'
+    'XT6s5v1Z<F}OkogF=z251`$@v-'
+    'CC9V_l_>OTg^1pvW!&nrF|LF9IQ?lT;@Pp~djCs9YFJuwnE)pC|IvyG95mEASkHA09fZ<HeH{{F@X~`Z+mvNqRvFmUKWCAv$b8pe'
+    '<69gS2DM|7}o-svJ5(5!gRu0o?7~|}CBX-Yu#rlLj#G<76RNi`XlzSeROfi5TW>Qhz)TYgoP|n&X8N1h4L0)CROPrOUpKZwxTq|~'
+    'v7K|z#sOeI_3~rf#QG7U**e=jjPvgh{C<6FTNv>obYPBGSWJ1sHc{6aW)-=vsP0GO7THn$V-'
+    'moYKt6^CfM>SDXpHCUrGTtP4FrDjrm3&eHL<K0dOij+V6kHav2crcZ!L94Ol599LLJH;wvSrcmW5LMTqbON=@DP`7<YYMsl3c{ug'
+    'eU?y^AmFPtV!7#yc|oa*W_E`TxAY?N#TYZgo_K;Q{(E1gqPss>zekqjeW2n&qAL+x|3;a*?1TS<)zeIe)EY1<+ZVRQ6*q23y(n$P'
+    'tOuWGJ`609A~si_QA_O{<42|u>F>v9lSgF@P7aF5Bl!p_~ZWh7?w3H;)D(IEO{hRo+SDGpelxBbJTkkM;UT>c=Cb%bhQ67eYgMam'
+    'xK4cS1=f7tT|5H?TaJw1E|{TDJ9M*gsD$&s)!0!(vShyRZ!-'
+    '9MyUO<3qySp!WDe_rSFk2Cr{*=+}{Gj#GOwAa$35%bU>yw&m2@q%{FHi0GcKv;z0ZvV1OE+68v&_i&1Emx*ToBk`Oq!@qL&`lZ>+'
+    '<sX{hN;;ecY#lT1fWLFrvb9i!md_gbH4(QR*JNn`GeTP8MU9@~Z3vP1`1f0445%c5e;cr)^yN#Ov-'
+    'g9Q~KwG1<l52GY#SDoOe`1i+SfE=t^WdQ6sD+XmS%uj*6r?lMdA6iAtpC#k<!;aDc&(w!j!<f;rG{ye1T?$%<%%mBED>zn0eVX<;'
+    'yw3^0=6V{q8XM0b~y;yO*Mff9Fw|`j<IdDm-ol}^!>q)7eCX}gZGE02k*}5FW~+S1g(PZ2-=uvh8q7tnUuxT-Xeyyb9>n9*TsQYQ'
+    '^u1juS6G_!1~V*f4(@mIQtU8I$|x8&SvooL{zaaLVNd9hWk}k;;xORZ<comJ9H{s&)yiVSfIkdsOG~CJggJ+YO#hh$Pgg@vXI_F0'
+    's@mryGkHOaLu%>k-{WFlcE}>A{|X9A{hotLj`Qs==&--k|L#_b3r-#TtU1fn{We9Ux(~24%lF?xaUF1lVS3v2xw4L0-'
+    '<*t6LXN(LO@vS)|f8T*J1@Tl&1dU4Wm_72_`wOqpoAMKGSypivt6y6G^R&2{pj2t}8{wEhL3u%2KmthTdHo(Kr)~m$Zn~CS+1T*j'
+    '?^Y>!It&k*GGLjpegokQk(kTwu}+IRoVBBMq})HF{hrMdoc%G~*v8??W4yFR|<RLBNE7%+~$w;PmjPKj`nLhvx^tBK)5}bF7f!I&'
+    'KAt+}k&_^kDf1dqEna9pG+`So5dyswjCI;3|Tk)nYxh8B~oFrWKe{q`sIMoeaV5`UIoYkB8^<c>nBI_aEPRo^5X{@cy%PadhM@o8'
+    '<w%Csf4$W-DfU`)28E%kW{=*_>ABT3rP!b7`~qb0}pX-'
+    '|iZMCQ7KH#Bw*yl<jfG==CFM?Gmpyvo@Zq^h1_^b8<I6q#nk&5oY#8BC$kTr^B)S%=t*bMxn2%z@#cbi$HILKQrRh(;Mx@a;6cIt'
+    'UUz@F)@;*TP$b<Ii=emptJ+KQC6|qgO>PQ*QRkdTc&omTVqH;62qP|*jsN%s|~6~$2O>T9uV&aB$0JV4cBRz(XBEOPcn$}NuEu-m'
+    'EN_$=JKh~+ZOn2S)%oh@MR9M(lIRaX*kI0eaZ3`*fw2#Vv+bIM0Hx`n1*;>!+jQwxYZv(S!N4ZVVsMGX>z4SgLG{20ndxTFP#2zk'
+    'ck9Z1tFRcjdet(A;r4&vx!U>I2*#QnJxMWk09*;tp#|3a3#KdLJ6L)#O9UTVrDjQGbd4)ZCH|-'
+    'c5~R>ZkS`Q)nc?Dz17yTGK2XVE5W!3+E-'
+    'SR+Q!j|IOZ!0k+ePc3wcT*L}m;tcCv|=J5Fa%nCk=Vs5XYzNvLTmUS#q*i>4Hr0mDq=0MhmZl5{+X^luVIAu43dm7jegzbM)evr2'
+    '0W;=7oUb)q4$2YY_-xq?c82IWJ&N^i}^*F`I+m6w>_yoQE}`HX-%9lOipkZ1+cstAEIs?+zFe>N$+YgPMDZq-'
+    'R#W!j}`v=G~m03t3QDeS9p^oZH5WWlqgRZQ){q7WZQBl$fw*f^U|0;f!_pyPC9T^pMQB%Au8-'
+    '`m*OAUAoQz?8T~eRU(IaY6A&$Ll@yHZ(G(6u7trq}(tK*~AO<3q?zAPx8Y+?95P`wGdojVK&38a}0RZ+cyMUkY&N6JwxD^ev(fBf'
+    '$?Yvb&*8TfIGe32B^~g#^QcsBduVS<emE(`2t*sLSe#&`ddNnC|solkAT$j9;eXcZ3(C!V0duh*276Qns9n0jQZibZ{)$LHn8Yw!'
+    '}G&%Dblkn)ks&iqzqm7`glzLb#g}EAN;g`adb|9+dsNEkn_RvwJ#857^9S~c9tu>$~vh1d3H|O;^3EK=>vGAY+JrQKL#%uU$1lbz'
+    'hI9EMT~#uCC-aNa%=YHVA4uUk}%j#Ru?oiP{>c`w>XWBy)>U|X8|@S@L=_6ZDCqgsM(|ZOHM^Q(d)ONtOxEOBP>J7c!omO`D#hAO'
+    'wz=%DwWfwd8l^5O0*OJ+_7vx22M}{z2R{<VG8xmPvTJKXqSI_*PB;u6~Pxy+HFp^d44ESuga^^QmVoto9es26lLXVkWKu20Wva6H'
+    'u1Bml}6>WVxi<#_L7B(;)kl}xG-{(P6YoqTAwSf-(|i!nje0u1RmOco#JIu@)E^E<=av1a*7V*{ys`ygWOg1Uae5o@vC3k9^-'
+    '5l*Irmowvp`{bY1Yt_hDieXq>&I_Upt8cXA$2XjEkZBq3JNbj?zU&DUjGESSnNM1x0USR#nqM6H*(Eqln{po&$g_{+~XmLIygFRT'
+    'I^p87RaU^C9Na)f7O;89ULabFu>*W|h|vjTk`H+W{k5jvKMmA*w4%b^AM;e@3$;*3$%l7OSTvY2W-=r8cyMAaF!^p{ow-qMq#Zl{'
+    '6WPXd=_R2}{LPmJ?YKBx{-'
+    'tVje(0i(O{XN0RgLuZHR+DHahdhgN$s#)wh{(o@wnJPLsygtj(g*s@j@F$?ZGN)qVA~Db;6k|9D%PPOy&u|v8C8j)sUQ5S>ov(Wb'
+    '^B0~E2D}Vj6khLno=s4Pu;WK~@jB39B@`4!F1N0C>>Q!hF(r}J+L9m;f4Wq1Ek|DV0kDym0#Y!Qy+%|7eu6>UFa2)kUglT}iu3KQ'
+    '7oBYhHqqx=vbt9g0DaxkMxzxhHH!%OTnIVls-0k`nT^fce_-d+6E421m~*19+H2Ppojohp-tOpW&zhDWgBaG99)bJ?#~`|OPoe4l'
+    'J(+FcLM&_F!@5Z~@2SQ#>0L`(&~#@Qb}HF`?Tuv$7#k97*tHD{UZcfe0kgHqKw;@*ei@d2JZe84(j+8&7Yg(APdD0PsP2krZzi?Y'
+    'ORL3e)_qA;$Eh=3(*cO?q?Xa@tg82wja!@g08fHhXpL5p@j`ouY2CvGwD2&G;~AaxWFY2}=8YK%^YY#j-'
+    'H#Z>VT!L}vDrZ&q2}A$6%=#@iWLU+`J_UBMG{|7ksx8cZi@n{#g&O4l3ki4Dk>7#BEngk-'
+    '?7jnRWk~;RBI;P0=fcIZA!LujZ?jsCJ?fAMZCp6HWgP=KhE5GtF-'
+    'L9N53&W_``k8*GKup^(?R<p0FFfU1u;T^D4NVa9;?B%jILic2$=)9+5*45Z3zHTPc9gy%aUC%AzXQVgm{$2%3oeLNdbDuLaTTrTA'
+    'h9mk_T-p4>4lx0=H|S9h%-UE2C}yujE@b@rs`L)0m5A(7gCJMfcaFwOG&Oo#7*LKxz2B4*uoXkZqhwq}!-'
+    'j&$93OQM*Ia!|mkX(D6tWKnybMK<|Mi(UD1jA4@dp>XX<=Rc0{Y=p~s092i7b~C!}AH!4$7scM_Y#KwTD8XQT*b-'
+    'ew1I0OR{!h=|P`!tlJIzyM#oVVUdwd3;0<4P3dL0J(`W6GNoZED5KKJJN!vgz|$K?f+LO8H2<h$=kk%Q)b&-'
+    'paMruOaa?+3if2HQmaJj7|Pu>AuLsl~8L3!}xp{2>99S8pKuLkwV4S&r`ZH+_4N4{M95<0U|>70K9#e0*&zM$uSd9*&WUfK>BEBX'
+    'jK=8U%lX;Pjp0ryRnG8uS)^#Q}3titP}Tl3n?J9HS4Vj{eO-bAA<@yxxBSFb{f3'
+)
+exec(compile(zlib.decompress(base64.b85decode(_PAYLOAD)), __file__, "exec"), globals(), globals())
